requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
dateparser==1.1.8
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
multidict==6.0.4
psycopg2-binary==2.9.6
pycryptodome==3.17
python-binance==1.0.17
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.5.5
requests==2.30.0
six==1.16.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==2.0.2
websocket==0.2.1
websocket-client==1.5.1
websockets==11.0.2
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
